+++ src/shared/api/apiRequest.ts
@@ -8,7 +8,7 @@ import { IAPIRequestConfig } from "@/shared/types";
  * @param {string} url - 요청할 엔드 포인트 예: '/api/v2/todo/{todoId}'
  * @param {'GET' | 'POST' | 'PATCH' | 'DELETE'} [method='GET'] - HTTP 요청 메서드 (기본 값은 GET)
  * @param {unknown} [data] - 요청에 포함할 데이터
- * @param {Record<string, string>} [headers={}] - //'Content-Type', 'multipart/form-data' 등 헤더 설정
+ * @param {Record<string, string>} [headers={}] - // 'Content-Type: multipart/form-data' 등 헤더 설정
  *
  * @returns {Promise<AxiosResponse>} - Axios 응답 객체
  */
@@ -27,7 +27,7 @@ const apiRequest = async ({
       headers,
     });
 
-    return response.data;
+    return response;
   } catch (error) {
     if (axios.isAxiosError(error) && error.response?.status === 401) {
       // 에러 코드에 따른 처리
